:truck: Component 별로 분리
